Code/requirements.txt
torch
torchvision
pillow
matplotlib
streamlit
pandas
scikit-learn
